Added jsp and Controller`s bodies;

+++ src/main/java/by/tms/bootstore/entity/order/BookInOrder.java
@@ -7,6 +7,6 @@ public class BookInOrder {
 
     private long id;
     private Book book;
-    private double countOfBook;
+    private int countOfBook;
 
 }
